refactor: update product mapping to use imgLink for externalId and cmId

+++ src/main/kotlin/io/github/havonte1/tcgwatcher/backend/adapter/out/persistence/repository/ProductRepositoryAdapter.kt
@@ -46,6 +46,8 @@ class ProductRepositoryAdapter(
 
     @Transactional
     override fun delete(product: Product) {
-        jpaRepository.findById(product.id).ifPresent { jpaRepository.delete(it) }
+        if (product.id != null) {
+            jpaRepository.findById(product.id).ifPresent { jpaRepository.delete(it) }
+        }
     }
 }
